fix: show full month name on meeting cards (Apr → April)

+++ web/src/components/MeetingListCard.tsx
@@ -19,7 +19,7 @@ function formatDayDate(dateStr: string): { day: string; monthDay: string } {
   const date = new Date(dateStr + 'T00:00:00')
   return {
     day: date.toLocaleDateString('en-US', { weekday: 'short' }),
-    monthDay: date.toLocaleDateString('en-US', { month: 'short', day: 'numeric' }),
+    monthDay: date.toLocaleDateString('en-US', { month: 'long', day: 'numeric' }),
   }
 }
